requirements.txt
boto3>=1.26.0
diagrams>=0.23.3
PyYAML>=6.0
click>=8.1.3
jinja2>=3.1.2
graphviz>=0.20.1
python-dateutil>=2.8.2
markdown>=3.4.3
weasyprint>=59.0
GitPython>=3.1.31
opa-python>=1.0.0
pydantic>=2.0.0

